Add JRuby support by removing ParseTree dependency.
