+++ lib/predicated/lib/predicated/from/callable_object.rb
@@ -35,6 +35,7 @@ require "predicated/from/ruby_code_string"
 module Predicated
   
   require_gem_version("ParseTree", "3.0.5", "parse_tree") if RUBY_VERSION < "1.9"
+  require "predicated/sexp_patch"
   
   class Predicate
 
@@ -64,44 +65,6 @@ module Predicated
         ruby_code_string = Ruby2Ruby.translate(temp_class, :serializable)    
         ruby_code_string.sub(/^def serializable\n  /, "").sub(/\nend$/, "")
       end
-    end
-    
-    #see http://gist.github.com/321038
-    # # Monkey-patch to have Ruby2Ruby#translate with r2r >= 1.2.3, from
-    # # http://seattlerb.rubyforge.org/svn/ruby2ruby/1.2.2/lib/ruby2ruby.rb
-    class ::Ruby2Ruby < ::SexpProcessor
-      def self.translate(klass_or_str, method = nil)
-        sexp = ParseTree.translate(klass_or_str, method)
-        unifier = Unifier.new
-        unifier.processors.each do |p|
-          p.unsupported.delete :cfunc # HACK
-        end
-        sexp = unifier.process(sexp)
-        self.new.process(sexp)
-      end
-      
-      #sconover - 7/2010 - monkey-patch
-      #{1=>2}=={1=>2}
-      #The right side was having its braces cut off because of 
-      #special handling of hashes within arglists within the seattlerb code.
-      #I tried to fork r2r and add a test, but a lot of other tests
-      #broke, and I just dont understand the test in ruby2ruby.
-      #So I'm emailing the author...
-      def process_hash(exp)
-        result = []
-        until exp.empty?
-          lhs = process(exp.shift)
-          rhs = exp.shift
-          t = rhs.first
-          rhs = process rhs
-          rhs = "(#{rhs})" unless [:lit, :str].include? t # TODO: verify better!
-
-          result << "#{lhs} => #{rhs}"
-        end
-
-        return "{ #{result.join(', ')} }"
-      end
-
     end
     
   end
