+++ lib/predicated/lib/predicated/from/ruby_code_string.rb
@@ -5,6 +5,7 @@ module Predicated
 
   require_gem_version("ruby_parser", "2.0.4")
   require_gem_version("ruby2ruby", "1.2.4")
+  require "predicated/sexp_patch"
 
   class Predicate
     def self.from_ruby_code_string(ruby_predicate_string, context=binding())
